* Data structures have been modified to use arrays instead of explicitly   named components. * Partial unrolling has been implemented in write_suN_headers.pl instead   of complete unrolling (unrolling parameters can be specified in this 	file): now the headers have constant size with NG. * Modifications to data structures have been propagated to all relevant 	source files, except for the TestProgram directory.
This is a huge update! Modifications need to be testes with the programs
in TestProgram dir...




git-svn-id: file:///Users/pica/Downloads/GitClone/svnrep/suN/HiRep@40 37a69121-b030-0410-a54a-fdca83ffee47

+++ Include/observables.h
@@ -37,6 +37,6 @@ void g5g3_correlator(double *out, suNf_spinor **qp);
 void g0g5g1_correlator(double *out, suNf_spinor **qp);
 void g0g5g2_correlator(double *out, suNf_spinor **qp);
 void g0g5g3_correlator(double *out, suNf_spinor **qp);
-void g5_g0g5_correlator_im(float *out, suNf_spinor **qp);
+void g5_g0g5_correlator_im(double *out, suNf_spinor **qp);
 
 #endif
